spmi: pmic-arb: Make irq_set_irq_wake() call conditional
Add a new device tree binding: qcom,not-wakeup, which can be used
to indicate that PMIC interrupts should not be treated as
wakeable.  If this property is not present, then
irq_set_irq_wake() will be called.  If it is present then
irq_set_irq_wake() will not be called.

This feature is convenient for bringup situations because it
breaks the dependency between spmi-pmic-arb and the MSM MPM
driver.

Change-Id: I44ac1221d92e13b4f9ab19571e95aa21d7ea10c2
Signed-off-by: David Collins <collinsd@codeaurora.org>