arm/dt: msm9625: Add qcom,not-wakeup property for SPMI arbiter
Add the qcom,not-wakeup flag to the SPMI arbiter device node so
that irq_set_irq_wake() is not called.  This ensures that PMIC
drivers may be tested before MPM, RPM-SMD, and SMD drivers are
validated on MSM9625.

This change should be reverted once the MPM driver is working
on target.

Change-Id: I539405902cebb4694c2f2a6e71d57e553ee5ab79
Signed-off-by: David Collins <collinsd@codeaurora.org>