arm/dt: msm9625: Switch from local to RPM control for all regulators
Switch to RPM control, via the rpm-regulator-smd driver, for all
PM8019 regulators used on MSM9625.  This is necessary so that
control is properly shared with other processors.

For certain shared regulators (S3, L10, and L12), the values
specified for regulator-min-microvolt should be lowered.  This is
necessary because the RPM can correctly aggregate requests from
multiple processors in order to guarantee that all voltage
requirements are met.  This is not possible with Linux directly
sets the voltage of these shared regulators via the
qpnp-regulator driver.

Select the qcom,init-voltage property values specified for
regulators which allow for a range of possible voltages based
upon which voltage is specified as default in the MDM9625 power
grid documentation.

Change-Id: I1317c4c3995718bafe79dc5ec588332943ff60ab
Signed-off-by: David Collins <collinsd@codeaurora.org>