msm: board-8064: Ensure that PM8921 GPIO 23 is only used on CDP and LIQUID
Add checks in the various APQ8064 board files so that PMIC PM8921
GPIO 23 is only utilized on CDP and LIQUID targets where it is
physically connected to a FET which enables the Atmel MXT
touchscreen.  This change makes it possible to use GPIO 23 for
other purposes on targets where it is not used to control the
touchscreen.

For CDP and LIQUID targets, msm-gpio-regulator "ext_ts_sw" is the
owner of PM8921 GPIO 23.  Therefore, make sure that this
regulator device is only registered on CDP and LIQUID targets.
Also modify the PMIC GPIO initialization tables so that GPIO 23
is only initialized for CDP and LIQUID.

Change-Id: Icb96bf3a096ceab747696b50f70901c6b6e1a8c7
CRs-Fixed: 406182
Signed-off-by: David Collins <collinsd@codeaurora.org>