msm: clock-local: Add PXO rates for LPASS clocks on 8660, 8960, 8930, 8064
LPASS clock are enabled out of reset on these SoCs, with PXO selected
as their undivided source. Because PXO rates are not present in the
frequency tables however, these clocks fail to handoff with reason
'HANDOFF_UNKNOWN_RATE'. As a result, clock_late_init() does not attempt
to disable these clocks in lateinit.

Add PXO rates to the LPASS frequency tables so that handoff succeeds
and, in lateinit, disables them if they are not used.

Change-Id: Ic2ef95b1401ff0d086825283dd19f3c8722e3feb
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>