msm: kgsl: Return 8064v1.1 chip ID for v1.2
Return the 8064v1.1 chip ID for v1.2 because there
appear to be no significant HW changes between the
revisions. This allows for the right SW workarounds
to be applied and fixes a major perf regression.

Note that v1.1 and v1.2 are currently considered the
exact same when looking at the bootloader chip ID, so
we have no way of knowing which it truly is.

Change-Id: Ia94b6b8af4d8378ee32fdde5f714b0cfd8e74f89
Signed-off-by: Kevin Matlage <kmatlage@codeaurora.org>