power: pm8921-charger: enable battery alarm lower comparator in suspend
Regardless of where the battery voltage is we should enable the
lower comparator while going to suspend. This will help us wakeup
when the battery goes below the low threshold while suspended.
Note that we are not impacting the hw clock switching algorithm.
The plan of disabling hw clock switching once we hit the low
threshold remains, it is only enabled when the battery gets above
upper threshold.

Change-Id: I8144e2b83748a96ab83ba14894bfa5bd59995895
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>